+++ fiddles/typescript/fiddle-0002-GenericService/source/ts/user.ts
@@ -1,5 +1,0 @@
-
-class User {
-    public name : string;
-    public password : string;
-}