+++ scripts/fiddle-refactor.sh
@@ -67,7 +67,7 @@ changeLogFile="../CHANGELOG.md";
             $(voidSubstr "${fiddleName}" "${newName}" "${altNewReadmeFile}";) || exit 91;
         fi
         ;;
-    'angular' | 'compass' | 'docker' | 'extjs5' | 'extjs6' | 'jquery' | 'meteor' | 'three' | 'php' | 'rxjs' | 'dojo' | 'chrome' | 'node' | 'tween' | 'typescript' | 'svg' )
+    'angular' | 'compass' | 'docker' | 'extjs5' | 'extjs6' | 'jquery' | 'meteor' | 'three' | 'php' | 'rxjs' | 'dojo' | 'chrome' | 'node' | 'tween' | 'svg' )
         if [[ -f "${newIndexFile}" ]]
         then
             $(voidSubstr "${fiddleName}" "${newName}" "${newIndexFile}";) || exit 91;